requirements.txt
Django==5.2.4
djangorestframework==3.15.2
Pillow==10.4.0
django-cors-headers==4.4.0
python-decouple==3.8
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-widget-tweaks==1.5.0
whitenoise==6.7.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
